Assert the release on the minimum footprint, and keep the legacy count 64-bit
Two changes.

Review feedback on #5540 (P2): legacyCycleAllocatedBytes was declared long,
but cn1LegacyBytesSinceGc is long long precisely because long is 32 bits on the
Windows LLP64 target. A collector falling more than 2GB of legacy allocation
behind would truncate the cycle count to a negative value and the quiet-cycle
test would read a furiously allocating app as idle -- reintroducing the
O(all pages) major sweep for the exact case the previous commit fixed. The
variable and the comparison are now 64-bit.

The probe's release assertion no longer reads the footprint at a chosen instant.
Reclamation is asynchronous and its LATENCY is load-dependent: the drop lands
about 2s after the ring is dropped on an idle macOS host, about 7s in an idle
Linux container, and had still not landed 15s in on a CI runner, where surefire
runs this probe alongside a dozen other forks and starves the collector. Two
successive attempts to fix that by waiting longer were both really measuring the
runner's load.

The claim under test is that the pages come back, not that they come back within
some number of seconds, so the app now tracks the LOWEST footprint seen at any
point after the warm-up's live set died and the assertion reads that. It cannot
be satisfied by a release that never happens, and it is immune to when the
release lands.

That this is the right reading is settled by the failing CI run's own numbers
rather than by argument: its table shows texture-after-small at base 32184KB,
i.e. the footprint DID fall to 32184KB against a 145981KB budget, while the
instant the old assertion sampled read 266784KB. The pages were always coming
back; the measurement was taken too early.

Also: the release settle no longer needs a long budget now that nothing depends
on it, so its cap drops from 60 rounds to 20. The probe runs in 34s on macOS,
down from 74s on CI.

Validation: full vm/tests suite 434 passed / 0 failures.

+++ vm/ByteCodeTranslator/src/cn1_globals.m
@@ -2269,7 +2269,11 @@ static long bibopCycleAllocatedBytes = 0;
 // partial page into every sweep is worst: heavy legacy churn driving the cycles
 // while a large BiBOP survivor set supplies the pages to walk -- precisely the
 // issue-5425 workload. GC-thread only, published at cycle begin.
-static long legacyCycleAllocatedBytes = 0;
+// long long, NOT long: cn1LegacyBytesSinceGc is long long for the same reason --
+// on the Windows LLP64 target long is 32 bits, so a collector that falls more
+// than 2GB of legacy allocation behind would truncate this to a negative value
+// and the quiet-cycle test below would read a furiously allocating app as idle.
+static long long legacyCycleAllocatedBytes = 0;
 static long bibopLastCycleOccupiedBytes = 0;
 static long bibopLastCycleLiveBytes = 0;
 static long bibopLastCycleReclaimedBytes = 0;
@@ -2398,8 +2402,8 @@ void cn1BibopBeginGcCycle(void) {
     // Same atomic-exchange idiom as the BiBOP reset above: a racing fetch_add
     // lands either before the swap (covered by the cycle that is starting) or
     // after it (charged to the next cycle) -- never dropped.
-    legacyCycleAllocatedBytes = (long)atomic_exchange_explicit(&cn1LegacyBytesSinceGc, 0,
-                                                               memory_order_acq_rel);
+    legacyCycleAllocatedBytes = atomic_exchange_explicit(&cn1LegacyBytesSinceGc, 0,
+                                                        memory_order_acq_rel);
     // Latch AFTER the counter: an allocator racing between the two exchanges
     // still sees the old latch and skips, so the fresh latch can never be
     // consumed by bytes just charged to the cycle that is starting.
@@ -3534,8 +3538,8 @@ static void cn1BibopSweep(CODENAME_ONE_THREAD_STATE) {
         // partial page in every sweep -- the O(all pages) regression issue 5425
         // fixed, reintroduced for exactly the workload that reported it.
         JAVA_BOOLEAN quiet =
-                (bibopCycleAllocatedBytes + legacyCycleAllocatedBytes)
-                        < CN1_BIBOP_MAJOR_SWEEP_QUIET_BYTES;
+                ((long long)bibopCycleAllocatedBytes + legacyCycleAllocatedBytes)
+                        < (long long)CN1_BIBOP_MAJOR_SWEEP_QUIET_BYTES;
         int major = atomic_load_explicit(&lowMemoryMode, memory_order_relaxed)
                 || quiet
                 || bibopCyclesSinceMajorSweep >= CN1_BIBOP_MAJOR_SWEEP_CYCLES;
